feat: initial pass of new preview components, generalize published + draft status components
