+++ packages/sanity/src/ui/index.ts
@@ -1,4 +1,6 @@
 export * from './button'
 export * from './dialog'
+export * from './draftStatus'
 export * from './menuItem'
+export * from './publishedStatus'
 export * from './tooltip'
